kill: handle TRAP signal
TRAP is a useful signal which when sent to a process currently
attached to a debugger causes the debugger to immediately break.
This is very helpful for programs running in some sort of run loop
where it may be difficult to set up a break point in an
appropriate location without modifying the source to include a
predicate condition.