tar: archive: improve fix for long names crashing
As requested, I resend my old patch for fixing the crashing while
archiving with names longer than 100 characters.

Last patch dealing with the issue was [1], and the old patch was [2]. The
code before this commit was not dealing correctly with multiple slashes,
but use of basename(3) and dirname(3) needed a temporary buffer because
otherwise we destroyed the path that was used later in several places.
This solution does not modifies the path and use pointer arithmetic to
solve the problem.

[1] https://lists.suckless.org/hackers/2412/19213.html
[2] https://lists.suckless.org/hackers/2402/19071.html

Co-authored-by: Roberto E. Vargas Caballer <k0ga@shike2.net>