tar: bug fix and compatibility improvements
1. A fix for a bug in the unarchive() function (causing metadata
loss when used on large tar files).  This bug is due to existing
code continuing to check for h->type == HARDLINK and SOFTLINK
(near end of function), when the entire header block has already
been overwritten by a call to eread() prior to the h->type checks.

2. Long (>=256 byte) file name compatibility: 'L' style long file
names were extremely simple to add, requiring a (net) addition of
a handful lines of code, and this patch supports both extracting
and creating L-style tar archives. Pax 'x' style long names are
more complex to parse, and this patch only supports extracting pax
'x' tars, but not creating them.

3. Command line argument compatibility improvements: 'c', 'x', 't'
args are accepted without needing a hyphen in front.  The '-p'
flag is also accepted but is no-op (as it is the normal behaviour
of sbase tar anyway, and allows sbase tar to be used in scripts
specifying this flag).  Directory tree member extraction is also
supported by this patch.

4. Handle tar archives with "." (current directory) entries.  Some
archives contain "." or "./" entries, causing error reports when
the current code tries to remove() the current dir.  I have added a
check in unarchive() to not perform remove() on encountering these
entries.