grep: Don't modify constant strings
The pattern pointer was assigned to a constant string and later it was
modified, which is UB. But even worse, as we were trusting the patlen
parameter received, patlen was meaningless after that assignment.

The addpattern() had many problems because it trusted the patlen
parameter, but in some places it used string functions that depend of
having a NUL character creating many problems when embedded NUL characters
were found in the pattern. As mandated by POSIX in [1]:

	The interfaces specified in POSIX.1-2017 do not permit the
	inclusion of a NUL character in an RE or in the string to be
	matched. If during the operation of a standard utility a NUL
	is included in the text designated to be matched, that NUL may
	designate the end of the text string for the purposes of matching.

so, the simples solution is just discard the patlen parameter and call
strlen() in addpattern() and use that as size of the pattern.

[1] https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap09.html

Co-authored-by: Santtu Lakkala <inz@inz.fi>