grep: Fix -x, simplify addpattern
Use grouping for -x to make \| / | work correctly; as a bonus beginning
and end of line anchors hold their special meaning inside a group,
removing the need for special handling of them.

Further, as this makes -w and -x only differ in anchors used, use the
same code for both.