bc: Implement return statements
When a return is found we have to go out of all the nested contexts
and _ideally_ that number should be in nested, but we are including
the function contexts here. Also, this makes clear that the
implementation of break was wrong, because we have to find the nesting
level of the previous loop to break it. Also, at this moment no
semantic check is done to vaalidate that we are in a loop or even
in a function, and it means that weird things can happen if we put
a return out of a function scope.