bc: Rewrite macros to implement break and return
The macros as implemented didn't allow correct tracking of the control
flow structures, and knowing he nesting level for return or break was
difficult. The tokens IF, WHILE and FOR were redirected to on terminal
symbols to be sure that the macro was already pushed before the rest
of the statement was parsed (including the global nested).