bc: Use dynamic memory for strings
Until now bc was using a clever mechanism that built all the strings
in a single buffer that makes very easy handling them because no
free was required. but it had a big problem because the size of the
space used in this buffer grew exponentially because new strings
were containing previous strings and all of them were living in memory
at the same time. This new mechanism implements a printf alike
and when strings are used as parameters it frees them, but it implies
that we cannot use anymore constants as parameters and in some specific
cases we have to free them manually.